+++ src/main/java/org/springframework/beans/factory/support/DefaultSingletonBeanRegistry.java
@@ -21,19 +21,21 @@ public class DefaultSingletonBeanRegistry implements SingletonBeanRegistry {
 
   private Map<String, Object> singletonObjects = new HashMap<>();
 
+  protected Map<String, Object> earlySingletonObjects = new HashMap<>();
+
   private final Map<String, DisposableBean> disposableBeans = new HashMap<>();
 
   @Override
   public Object getSingleton(String beanName) {
-    return singletonObjects.get(beanName);
+    Object bean = singletonObjects.get(beanName);
+    if (bean == null) {
+      bean = earlySingletonObjects.get(beanName);
+    }
+    return bean;
   }
 
   @Override
   public void registerSingleton(String beanName, Object singletonObject) {
-    singletonObjects.put(beanName, singletonObject);
-  }
-
-  protected void addSingleton(String beanName, Object singletonObject) {
     singletonObjects.put(beanName, singletonObject);
   }